refactor:made code review changes from @dkibet
