+++ packages/esm-billing-app/src/constants.ts
@@ -1,3 +1,6 @@
 export const PENDING_PAYMENT_STATUS = 'PENDING';
 export const EXEMPTED_PAYMENT_STATUS = 'EXEMPTED';
 export const spaBasePath = `${window.spaBase}/home`;
+
+// Size in MegaBits
+export const MAX_ALLOWED_FILE_SIZE = 2097152;
